Created viewer for projects page
Started making the layout for the projects page. It’s just a static
page for now and the code still needs to be cleaned up quite a bit, but
should give a nice introduction to how this page can look. Next step is
to make a view where users can add their project info, and have it
dynamically update the list of projects. Let me know what you all think
of it so far.

As a side note, this was pulled from master and is being committed to
projects, so it should have all of the code that is currently
implemented in the master branch